requirements.txt
Django==2.0.4
django-froala-editor==2.5.1
django-material==0.12.3
freeze==1.0.10
mysqlclient==1.3.10
olefile==0.44
Pillow==4.1.1
pytz==2018.3
six==1.10.0
